Move TracingSpoolingManager to main
It should apply to all SpoolingManager implementations
